requirements.txt
openai>=1.12.0
pypdf>=4.0.0
pydantic>=2.6.0
gradio>=4.19.0
python-dotenv>=1.0.0
pytest>=8.0.0
anthropic>=0.18.1